Add pre-commit

+++ roles/fluentd/templates/test2/test2.sh
@@ -1,4 +1,14 @@
 #!/usr/bin/env bash
 
-pkill -f fluentd
+set -x
+set -e
+set -u
+
+systemctl stop fluentd
+
+if pgrep fluetnd; then
+    pkill -f fluentd
+fi
+
+/usr/sbin/fluent-gem install fluent-plugin-detect-exceptions
 fluentd --verbose --config ./fluentd.conf
